+++ src/test/compile-fail/issue-4968.rs
@@ -10,9 +10,9 @@
 
 // Regression test for issue #4968
 
-const A: (int,int) = (4,2);
+const A: (isize,isize) = (4,2);
 fn main() {
     match 42 { A => () }
-    //~^ ERROR mismatched types: expected `_`, found `(int, int)`
+    //~^ ERROR mismatched types: expected `_`, found `(isize, isize)`
     //         (expected integral variable, found tuple)
 }
